tidy up backend | Hurray!

+++ backend/controllers/product.controller.js
@@ -52,11 +52,15 @@ export const updateProduct = async (req, res) => {
 export const deleteProduct = async (req, res) => {
     const { id } = req.params;
 
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({success: false, message: "Invalid product id"});
+    }
+    
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({success: true, message: "Product Deleted"});
     } catch (error) {
         onsole.log("error in deleting products: ", error.message);
-        res.status(404).json({success: false, message: "Product not found"});
+        res.status(500).json({success: false, message: "Server Error"});
     }
 };
